vision_proxy/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
google-cloud-vision
python-multipart